requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
tenacity==8.2.3
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3
scikit-learn==1.3.2
streamlit==1.29.0
openpyxl==3.1.2
haversine==2.8.0
